Feature: WhatIf (#12)
* Implement `-whatif` for `backup`, `restore`, and `cleanup`

* Update README to explain the `-whatif` flag

+++ commandClean.go
@@ -158,6 +158,13 @@ func (c *cmdCleanup) handleCleanupQueue(id int) {
 			// Remove the entry from the backup, and from the backup index.
 			glog.V(1).Infof("[Cleanup-%d] Remove file '%s' from backup ...",
 				id, m.Entry.RelPath)
+
+			if c.args.whatIf {
+				numSuccessful++
+				fmt.Println(m.Entry.RelPath)
+				continue
+			}
+
 			if err := c.archive.Delete(m.Entry); nil != err {
 				numFailed++
 				glog.Errorf("[Cleanup-%d] Removal of file '%s' failed: %v",
@@ -171,6 +178,13 @@ func (c *cmdCleanup) handleCleanupQueue(id int) {
 			// Remove the file from the local file system.
 			glog.V(1).Infof("[Cleanup-%d] Remove local file '%s' ...",
 				id, m.relPath)
+
+			if c.args.whatIf {
+				numSuccessful++
+				fmt.Println(m.relPath)
+				continue
+			}
+
 			if err := os.Remove(m.absolutePath); nil != err {
 				numFailed++
 				glog.Errorf("[Cleanup-%d] Removal of local file '%s' failed: %v",
